Consolidate dashboards listing page into a shared web-common component
Extract the project dashboards-listing UI into
`web-common/src/features/preview-mode/DashboardsPage.svelte`. Both
the local route (`/dashboards`) and the cloud editor dev-preview
route (`/-/edit/dashboards`) now thinly wrap it; only the link
builder differs. The web-local-only `DashboardList.svelte` shim is
deleted.

First step in unifying the local app's preview surface with the
cloud editor's `/-/edit/...` chrome.
